fix warning

+++ ffplayer/decoder_video.cc
@@ -45,7 +45,8 @@ int VideoDecoder::DecodeThread() {
     auto duration = (frame_rate.num && frame_rate.den
                          ? av_q2d(AVRational{frame_rate.den, frame_rate.num})
                          : 0);
-    auto pts = (frame->pts == AV_NOPTS_VALUE) ? NAN : frame->pts * av_q2d(tb);
+    auto pts =
+        (frame->pts == AV_NOPTS_VALUE) ? NAN : double(frame->pts) * av_q2d(tb);
 
     ret = video_render_->PushFrame(frame, pts, duration, pkt_serial);
     av_frame_unref(frame);
@@ -60,8 +61,7 @@ int VideoDecoder::DecodeThread() {
 VideoDecoder::VideoDecoder(unique_ptr_d<AVCodecContext> codecContext,
                            std::unique_ptr<DecodeParams> decodeParams,
                            std::shared_ptr<VideoRenderBase> render)
-    : Decoder(std::move(codecContext),
-              std::move(decodeParams)),
+    : Decoder(std::move(codecContext), std::move(decodeParams)),
       video_render_(std::move(render)) {
   Start();
 }
